zuverlässigere flankenerkennung
git-svn-id: svn+ssh://blue.3dallee.de/fhtw/trunk@63 6aeb069d-d5e0-465d-890a-ccf59a9091f3

+++ src/AudioSink_DirectSound.c
@@ -10,7 +10,7 @@
 #include <time.h>
 
 #define numbuffers 8
-#define BUFFER_SIZE ((int)ag_SAMPLERATE*8)
+#define BUFFER_SIZE (ag_SAMPLERATE*8)
 
 extern void AudioSink_OnData(ComplexPackage);
